docs: change the doc to highlight the allowed multicast addresses
Bugzilla ID: 802

The ipv4_multicast example does not work with any multicast IPs.
Only a selected few IPs are allowed to send pakcets. These IPs
are listed in an mcast_group_table array along with their respective
port masks. A normal user would not know about this behaviour since
there is no mention of it in the docs

Added the mcast_group_table in the docs so user would know which
IPs are allowed.

Signed-off-by: huzaifa.rahman <huzaifa.rahman@emumba.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>