doc: announce behavior change for device NUMA node field
The dev->device.numa_node field is set by the bus driver for each device
it manages to indicate on which NUMA node this device lies.

When this information is unknown, the assigned value is not consistent
across the bus drivers. In v22.11, the value will be set to -1 in
this case.

Link: https://patches.dpdk.org/project/dpdk/patch/20211026090610.10823-1-houssem.bouhlel@6wind.com/

Signed-off-by: Olivier Matz <olivier.matz@6wind.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>