vhost: stop using mempool for IOTLB
A mempool consumes 3 memzones (with the default ring mempool driver).
Default DPDK configuration allows RTE_MAX_MEMZONE (2560) memzones.

Assuming there is no other memzones (well, there is always other
memzones in a DPDK application), that means that we can have a maximum
of 853 mempools.

The IOTLB so far was requesting a mempool per vq, which means that at
the maximum, the vhost library can request mempools for 426 qps.

This limit was recently reached on big systems with a lot of virtio
ports (and multiqueue in use).

While the limit on mempool count could be something we need to look at
the DPDK project level, there is no reason to use mempools with the
IOTLB code.

The IOTLB cache entries do not need to be DMA-able and are only used by
the current process (in multiprocess context).
Getting/putting objects from/in the mempool is always associated with
some other locks, so some level of contention is already present.

We can convert to a malloc'd pool with objects put in a free list
protected by a spinlock.

Signed-off-by: David Marchand <david.marchand@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>