devtools: add trace function check in checkpatch
This patch adds a check in checkpatch tool, to check if trace
function is added in any new function added in ethdev library.

It uses the existing build_map_changes() function and version.map
file to create a list of newly added functions. The definition of
these functions are checked if they contain trace calls or not.
The checkpatch return error if the trace calls are not present.

Cc: tduszynski@marvell.com

Signed-off-by: Ankur Dwivedi <adwivedi@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>