devtools: eol abi as a separate function
Developer tools associated with abi are maintained with as part of
developer tooling, eal abi headers are maintained with eal, abi build
scripts are maintained with the build system and abi policy and version
documents along with rest of the documentation.

Major change is that individual components maintainers become
responsible for ensuring correctness of their map file(s).

Signed-off-by: Ray Kinsella <mdr@ashroe.eu>
Signed-off-by: 0-day Robot <robot@bytheb.org>